+++ glyph-brush-layout/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.1.5
 * Add `GlyphPositioner::recalculate_glyphs` with a default unoptimised implementation. Custom layouts won't be broken by this change, but _will_ need to implement the new function to provide optimised behaviour.
 * Optimise built-in layout's recalculate_glyphs for screen position changes with `GlyphChange::Geometry`.
 * Optimise built-in layout's recalculate_glyphs for single color changes with `GlyphChange::Color`.